Rename GUI class
