+++ qupath-gui-fx/src/main/java/qupath/lib/gui/ExtensionManager.java
@@ -278,295 +278,6 @@ public class ExtensionManager {
 		refreshExtensions(true);
 	}
 
-
-	/**
-	 * Borderpane that displays extensions, with options to remove,
-	 * open containing folder, update, where possible.
-	 */
-	static class ExtensionManagerController extends BorderPane {
-
-		private static final Logger logger = LoggerFactory.getLogger(QuPathGUI.class);
-
-		@FXML
-		private ListView<QuPathExtension> listView;
-
-		@FXML
-		private Button addBtn;
-
-		@FXML
-		private Button rmBtn;
-		@FXML
-		private Button disableBtn;
-		@FXML
-		private Button submitBtn;
-		@FXML
-		private Button openExtensionDirBtn;
-
-		@FXML
-		private Button updateBtn;
-
-		@FXML
-		private VBox topVBox;
-		@FXML
-		private HBox addHBox;
-
-		@FXML
-		private TextField ownerTextArea;
-		@FXML
-		private TextField repoTextArea;
-
-		/**
-		 * Create an instance of the ExtensionManager UI pane.
-		 * @return A BorderPane subclass.
-		 * @throws IOException If FXML or resources can't be found.
-		 */
-		static ExtensionManagerController createInstance() throws IOException {
-			return new ExtensionManagerController();
-		}
-
-		private ExtensionManagerController() throws IOException {
-			var loader = new FXMLLoader(ExtensionManagerController.class.getResource("ExtensionManager.fxml"));
-			loader.setController(this);
-			loader.setRoot(this);
-			loader.setResources(ResourceBundle.getBundle("qupath/lib/gui/localization/qupath-gui-strings"));
-			loader.load();
-		}
-
-
-		@FXML
-		private void initialize() {
-			this.setOnDragDropped(QuPathGUI.getInstance().getDefaultDragDropListener());
-			ExtensionManager extensionManager = QuPathGUI.getInstance().getExtensionManager();
-			ObservableMap<Class<? extends QuPathExtension>, QuPathExtension> extensions = extensionManager.getLoadedExtensions();
-			extensions.addListener((MapChangeListener<Class<? extends QuPathExtension>, QuPathExtension>) c -> {
-				if (c.wasAdded()) {
-					listView.getItems().add(c.getValueAdded());
-				}
-				if (c.wasRemoved()) {
-					listView.getItems().remove(c.getValueRemoved());
-				}
-			});
-
-			openExtensionDirBtn.disableProperty().bind(
-					UserDirectoryManager.getInstance().userDirectoryProperty().isNull());
-
-			var items = listView.getItems();
-			items.addAll(
-					extensionManager.getLoadedExtensions().values()
-							.stream()
-							.sorted(Comparator.comparing(QuPathExtension::getName))
-							.toList());
-			listView.setCellFactory(param -> new ExtensionListCell(param));
-			ownerTextArea.addEventHandler(KeyEvent.KEY_RELEASED, e -> {
-				if (e.getCode() == KeyCode.ENTER) {
-					submitAdd();
-				}
-				if (e.getCode() == KeyCode.ESCAPE) {
-					cancelAdd();
-				}
-			});
-			repoTextArea.addEventHandler(KeyEvent.KEY_RELEASED, e -> {
-				if (e.getCode() == KeyCode.ENTER) {
-					submitAdd();
-				}
-				if (e.getCode() == KeyCode.ESCAPE) {
-					cancelAdd();
-				}
-			});
-		}
-
-		@FXML
-		private void addExtension() {
-			if (!topVBox.getChildren().contains(addHBox)) {
-				topVBox.getChildren().add(addHBox);
-			}
-		}
-
-		@FXML
-		private void submitAdd() {
-			var repo = GitHubProject.GitHubRepo.create("tmp", ownerTextArea.getText(), repoTextArea.getText());
-			try {
-				DragDropImportListener.askToDownload(repo);
-			} catch (URISyntaxException | IOException | InterruptedException e) {
-				Dialogs.showErrorNotification(QuPathResources.getString("ExtensionManager.unableToDownload"), e);
-			}
-			cancelAdd();
-		}
-
-		@FXML
-		private void cancelAdd() {
-			ownerTextArea.clear();
-			repoTextArea.clear();
-			topVBox.getChildren().remove(addHBox);
-		}
-
-		@FXML
-		private void openExtensionDir() {
-			var dir = ExtensionClassLoader.getInstance().getExtensionDirectory();
-			if (dir != null) {
-				GuiTools.browseDirectory(dir.toFile());
-			} else {
-				Dialogs.showErrorNotification(
-						QuPathResources.getString("ExtensionManager"),
-						QuPathResources.getString("ExtensionManager.noExtensionDirectorySet"));
-			}
-		}
-
-		private static void removeExtension(QuPathExtension extension) {
-			if (extension == null) {
-				logger.info("No extension selected, so none can be removed");
-				return;
-			}
-			if (!Dialogs.showYesNoDialog(
-					QuPathResources.getString("ExtensionManager"),
-					String.format(QuPathResources.getString("ExtensionManager.confirmRemoveExtension"), extension.getName()))) {
-				return;
-			}
-			try {
-				var url = extension.getClass().getProtectionDomain().getCodeSource().getLocation();
-				logger.info("Removing extension: {}", url);
-				new File(url.toURI().getPath()).delete();
-				Dialogs.showInfoNotification(
-						QuPathResources.getString("ExtensionManager"),
-						String.format(QuPathResources.getString("ExtensionManager.extensionRemoved"), url));
-			} catch (URISyntaxException e) {
-				logger.error("Exception removing extension: " + extension, e);
-			}
-		}
-
-		private static void updateExtension(QuPathExtension extension) {
-			if (!(extension instanceof GitHubProject project)) {
-				Dialogs.showWarningNotification(QuPathResources.getString("ExtensionManager"),
-						QuPathResources.getString("ExtensionManager.unableToCheckForUpdates"));
-				return;
-			}
-			var version = extension.getVersion();
-			try {
-				var release = UpdateChecker.checkForUpdate(project.getRepository());
-				if (release != null && release.getVersion() != Version.UNKNOWN && version.compareTo(release.getVersion()) < 0) {
-					logger.info("Found newer release for {} ({} -> {})", project.getRepository().getName(), version, release.getVersion());
-					DragDropImportListener.askToDownload(project.getRepository());
-				} else if (release != null) {
-					Dialogs.showInfoNotification(
-							QuPathResources.getString("ExtensionManager"),
-							String.format(
-									QuPathResources.getString("ExtensionManager.noNewerRelease"),
-									project.getRepository().getName(), version, release.getVersion())
-					);
-				}
-			} catch (IOException | URISyntaxException | InterruptedException e) {
-				Dialogs.showErrorNotification(QuPathResources.getString("ExtensionManager.unableToUpdate"), e);
-			}
-		}
-
-		/**
-		 * Controller class for extension list cells
-		 */
-		static class ExtensionListCell extends ListCell<QuPathExtension> {
-
-			public ExtensionListCell(ListView<QuPathExtension> listView) {
-				super();
-			}
-
-			@Override
-			public void updateItem(QuPathExtension item, boolean empty) {
-				super.updateItem(item, empty);
-				if (empty || item == null) {
-					setText(null);
-					setGraphic(null);
-					return;
-				}
-				if (!(item instanceof GitHubProject)) {
-					setMouseTransparent(true);
-					setFocusTraversable(false);
-					setDisable(true);
-				}
-				ExtensionListCellVBox vbox = new ExtensionListCellVBox(item);
-
-				vbox.rmBtn.setGraphic(IconFactory.createNode(8, 8, IconFactory.PathIcons.MINUS));
-				vbox.updateBtn.setGraphic(IconFactory.createNode(8, 8, IconFactory.PathIcons.REFRESH));
-				vbox.nameText.setText(item.getName());
-				vbox.versionText.setText(item.getVersion().toString());
-				vbox.descriptionText.setText(WordUtils.wrap(item.getDescription(), 80));
-
-				setGraphic(vbox);
-				var contextMenu = new ContextMenu();
-				contextMenu.getItems().add(ActionTools.createMenuItem(
-						ActionTools.createAction(() -> openContainingFolder(item),
-								QuPathResources.getString("ExtensionManager.openContainingFolder"))));
-				contextMenu.getItems().add(ActionTools.createMenuItem(
-						ActionTools.createAction(() -> updateExtension(item),
-								QuPathResources.getString("ExtensionManager.updateExtension"))));
-				contextMenu.getItems().add(ActionTools.createMenuItem(
-						ActionTools.createAction(() -> removeExtension(item),
-								QuPathResources.getString("ExtensionManager.removeExtension"))));
-				this.setContextMenu(contextMenu);
-
-				var tooltipText = item.getName() + "\n" + QuPathResources.getString("ExtensionManager.doubleClick");
-				Tooltip.install(this, new Tooltip(tooltipText));
-				setOnMouseClicked(event -> {
-					if (event.getClickCount() == 2 && event.getButton() == MouseButton.PRIMARY) {
-						if (item instanceof GitHubProject) { // this should always be true anyway...
-							String url = ((GitHubProject)item).getRepository().getUrlString();
-							try {
-								logger.info("Trying to open URL {}", url);
-								GuiTools.browseURI(new URI(url));
-							} catch (URISyntaxException e) {
-								Dialogs.showErrorNotification(
-										QuPathResources.getString("ExtensionManager.unableToOpenGitHubURL") + url,
-										e);
-							}
-						}
-					}
-				});
-			}
-
-			private void openContainingFolder(QuPathExtension extension) {
-				var url = extension.getClass().getProtectionDomain().getCodeSource().getLocation();
-				try {
-					GuiTools.browseDirectory(new File(url.toURI()));
-				} catch (URISyntaxException e) {
-					logger.error("Unable to open directory {}", url);
-				}
-			}
-		}
-
-		/**
-		 * Simple class that just loads the FXML for a list cell
-		 */
-		static class ExtensionListCellVBox extends VBox {
-			private final QuPathExtension extension;
-			@FXML
-			private Button rmBtn;
-			@FXML
-			private Button updateBtn;
-			@FXML
-			private Text nameText, versionText, descriptionText;
-			ExtensionListCellVBox(QuPathExtension extension) {
-				this.extension = extension;
-				var loader = new FXMLLoader(ExtensionManagerController.class.getResource("ExtensionListCellVBox.fxml"));
-				loader.setController(this);
-				loader.setRoot(this);
-				loader.setResources(ResourceBundle.getBundle("qupath/lib/gui/localization/qupath-gui-strings"));
-				try {
-					loader.load();
-				} catch (IOException e) {
-					throw new RuntimeException(e);
-				}
-			}
-
-			@FXML
-			private void updateExtension() {
-				ExtensionManagerController.updateExtension(this.extension);
-			}
-			@FXML
-			private void removeExtension() {
-				ExtensionManagerController.removeExtension(this.extension);
-			}
-		}
-
-	}
-
 	/**
 	 * Create a BorderPane for displaying/managing extensions.
 	 * @return A BorderPane for displaying and managing QuPath extensions.
