+++ qupath-gui-fx/src/main/resources/qupath/lib/gui/localization/qupath-gui-strings.properties
@@ -907,30 +907,30 @@ DragDrop.couldNotHandleFile = Sorry, I couldn't figure out what to do with %s
 
 
 ## Extension manager GUI pane
-ExtensionManager = Extension manager
-ExtensionManager.unableToOpen = Unable to open extension manager
-ExtensionManager.unableToDownload = Unable to download extension from GitHub
-ExtensionManager.confirmRemoveExtension = Are you sure you want to remove %s?
-ExtensionManager.extensionRemoved = Extension removed %s\nYou need to restart QuPath to see the changes reflected.
-ExtensionManager.unableToCheckForUpdates = Unable to check for updates
-ExtensionManager.installExtensionFromGithub = Would you like to install the extension %s from GitHub?
-ExtensionManager.successfullyDownloaded = Successfully downloaded %s
-ExtensionManager.noNewerRelease = No newer release for %s (%s versus %s)
-ExtensionManager.unableToUpdate = Unable to update extension
-ExtensionManager.openContainingFolder = Open containing folder
-ExtensionManager.updateExtension = Update Extension
-ExtensionManager.removeExtension = Remove extension
-ExtensionManager.doubleClick = Double-click to view website
-ExtensionManager.unableToOpenGitHubURL = Unable to open GitHub URL:
-ExtensionManager.add = Add extension
-ExtensionManager.addTooltip = Install extension by specifying a GitHub username and repository 
-ExtensionManager.owner = Owner
-ExtensionManager.ownerTooltip = The username that the extension can be found under. e.g., qupath
-ExtensionManager.repository = Repository
-ExtensionManager.repositoryTooltip = The repository that the extension can be found in.\ne.g., qupath-extension-stardist
-ExtensionManager.noExtensionDirectorySet = User directory has not been set. Install an extension to set a user directory
-ExtensionManager.openExtensionDirectory = Open user directory containing extensions and other files
-ExtensionManager.accept = Accept
-ExtensionManager.acceptTooltip = Attempt to install the extension using the specified Owner and Repository
-ExtensionManager.cancel = Cancel
-ExtensionManager.cancelTooltip = Clear the Owner and Repository fields without installing an extension
+ExtensionControlPane = Extension manager
+ExtensionControlPane.unableToOpen = Unable to open extension manager
+ExtensionControlPane.unableToDownload = Unable to download extension from GitHub
+ExtensionControlPane.confirmRemoveExtension = Are you sure you want to remove %s?
+ExtensionControlPane.extensionRemoved = Extension removed %s\nYou need to restart QuPath to see the changes reflected.
+ExtensionControlPane.unableToCheckForUpdates = Unable to check for updates
+ExtensionControlPane.installExtensionFromGithub = Would you like to install the extension %s from GitHub?
+ExtensionControlPane.successfullyDownloaded = Successfully downloaded %s
+ExtensionControlPane.noNewerRelease = No newer release for %s (%s versus %s)
+ExtensionControlPane.unableToUpdate = Unable to update extension
+ExtensionControlPane.openContainingFolder = Open containing folder
+ExtensionControlPane.updateExtension = Update Extension
+ExtensionControlPane.removeExtension = Remove extension
+ExtensionControlPane.doubleClick = Double-click to view website
+ExtensionControlPane.unableToOpenGitHubURL = Unable to open GitHub URL:
+ExtensionControlPane.add = Add extension
+ExtensionControlPane.addTooltip = Install extension by specifying a GitHub username and repository 
+ExtensionControlPane.owner = Owner
+ExtensionControlPane.ownerTooltip = The username that the extension can be found under. e.g., qupath
+ExtensionControlPane.repository = Repository
+ExtensionControlPane.repositoryTooltip = The repository that the extension can be found in.\ne.g., qupath-extension-stardist
+ExtensionControlPane.noExtensionDirectorySet = User directory has not been set. Install an extension to set a user directory
+ExtensionControlPane.openExtensionDirectory = Open user directory containing extensions and other files
+ExtensionControlPane.accept = Accept
+ExtensionControlPane.acceptTooltip = Attempt to install the extension using the specified Owner and Repository
+ExtensionControlPane.cancel = Cancel
+ExtensionControlPane.cancelTooltip = Clear the Owner and Repository fields without installing an extension
